ops(heroku):speed up build

+++ scripts/reduce-slug-size-heroku.sh
@@ -1,5 +1,6 @@
 #!/bin/bash
 [ -z "${LUMIUM_HEROKU_BUILD}" ] && exit 0;
+[ -z "${ENABLE_REDUCE_SLUG_SIZE_HEROKU}" ] && exit 0;
 echo "cleaning up..."
 rm -rf lumium-renderer/target
 rm -rf clang+llvm-14.0.0-x86_64-linux-gnu-ubuntu-18.04
